binaural_stimuli_generator/requirements.txt
alabaster==0.7.16
Babel==2.15.0
certifi==2024.6.2
cffi==1.16.0
cftime==1.6.3
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
docutils==0.21.2
fonttools==4.53.0
idna==3.7
imagesize==1.4.1
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
miniaudio==1.59
netCDF4==1.6.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
py3dti==0.3.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
pysofaconventions==0.1.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-posthocs==0.9.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
snowballstemmer==2.2.0
spatialaudiometrics==0.0.4
statsmodels==0.14.2
tzdata==2024.1
urllib3==2.2.1
